bot/requirements.txt
python-telegram-bot==22.4
openai==1.107.2
